Refactor parser to deal with obvious incomplete commands
Place parser and storage into io package for better readability and
categorization
