+++ src/main/java/com/jacob/duke/command/ByeCommand.java
@@ -1,7 +1,7 @@
 package main.java.com.jacob.duke.command;
 
 import main.java.com.jacob.duke.DukeList;
-import main.java.com.jacob.duke.Storage;
+import main.java.com.jacob.duke.io.Storage;
 import main.java.com.jacob.duke.Ui;
