+++ src/main/java/com/jacob/duke/command/Command.java
@@ -2,7 +2,7 @@ package main.java.com.jacob.duke.command;
 
 import main.java.com.jacob.duke.DukeException;
 import main.java.com.jacob.duke.DukeList;
-import main.java.com.jacob.duke.Storage;
+import main.java.com.jacob.duke.io.Storage;
 import main.java.com.jacob.duke.Ui;
