+++ src/main/java/com/jacob/duke/command/NoteCommand.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import main.java.com.jacob.duke.DukeException;
 import main.java.com.jacob.duke.DukeList;
-import main.java.com.jacob.duke.Storage;
+import main.java.com.jacob.duke.io.Storage;
 import main.java.com.jacob.duke.Ui;
 import main.java.com.jacob.duke.note.Note;
 
@@ -16,13 +16,8 @@ public class NoteCommand implements Command {
         this.fullCommand = fullCommand;
     }
     @Override
-    public String execute(Ui ui, DukeList dukeList, Storage storage) throws DukeException {
+    public String execute(Ui ui, DukeList dukeList, Storage storage) {
         int breakpoint = fullCommand.indexOf("?");
-        if (fullCommand.length() <= "note".length()) {
-            throw new DukeException("A note cannot be empty!");
-        } else if (breakpoint == -1) {
-            throw new DukeException("Hey, a note must have a corresponding question!");
-        }
         //before break is question
         String question = fullCommand.substring("note".length() + 1, breakpoint + 1);
         //after break is answer till the end
@@ -30,6 +25,7 @@ public class NoteCommand implements Command {
         Note theNote = new Note(question, answer);
         List<Note> noteList = dukeList.getNoteList();
         noteList.add(theNote);
+
         storage.appendTextToNotes(theNote.convertToFileFormat());
         return ui.showNewNoteAdded(theNote.getCurrentStatus(), noteList);
     }
