+++ src/main/java/com/jacob/duke/command/PrintListCommand.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import main.java.com.jacob.duke.DukeException;
 import main.java.com.jacob.duke.DukeList;
-import main.java.com.jacob.duke.Storage;
+import main.java.com.jacob.duke.io.Storage;
 import main.java.com.jacob.duke.Ui;
 import main.java.com.jacob.duke.task.Task;
