+++ src/main/java/com/jacob/duke/command/PrintNoteListCommand.java
@@ -3,7 +3,7 @@ import java.util.List;
 
 import main.java.com.jacob.duke.DukeException;
 import main.java.com.jacob.duke.DukeList;
-import main.java.com.jacob.duke.Storage;
+import main.java.com.jacob.duke.io.Storage;
 import main.java.com.jacob.duke.Ui;
 import main.java.com.jacob.duke.note.Note;
