+++ src/main/java/com/jacob/duke/command/TodoCommand.java
@@ -2,9 +2,8 @@ package main.java.com.jacob.duke.command;
 
 import java.util.List;
 
-import main.java.com.jacob.duke.DukeException;
 import main.java.com.jacob.duke.DukeList;
-import main.java.com.jacob.duke.Storage;
+import main.java.com.jacob.duke.io.Storage;
 import main.java.com.jacob.duke.Ui;
 import main.java.com.jacob.duke.task.Task;
 import main.java.com.jacob.duke.task.Todo;
@@ -28,17 +27,13 @@ public class TodoCommand implements Command {
      * @param ui UI object to deal with program output.
      * @param dukeList Task List Representation.
      * @param storage Storage object to deal with interfacing with file system.
-     * @throws DukeException In case there are internal errors.
      */
     @Override
-    public String execute(Ui ui, DukeList dukeList, Storage storage) throws DukeException {
+    public String execute(Ui ui, DukeList dukeList, Storage storage) {
         List<Task> taskList = dukeList.getTaskList();
         String description = inputCommand.substring(4);
 
-        if (description.equals("")) {
-            throw new DukeException(" ☹ OOPS!!! The description of a todo cannot be empty.");
-        }
-        Task theTodo = new Todo(inputCommand.substring(4 + 1));
+        Task theTodo = new Todo(description.trim());
         taskList.add(theTodo);
 
         storage.appendTextToTasks(theTodo.convertToFile());
